requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.5
django-rest-framework==0.1.0
django-silk==5.4.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
gprof2dot==2025.4.14
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mysqlclient==2.2.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
